# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Worker Safety Monitoring API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalWorkers: number;
  activeWorkers: number;
  totalViolations: number;
  activeAlerts: number;
  safetyScore: number;
  camerasOnline: number;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Alert {
  id: number;
  type: string;
  severity: AlertSeverity;
  message: string;
  /** @nullable */
  workerId?: number | null;
  zone: string;
  timestamp: string;
  resolved: boolean;
}

export interface ActivityDataPoint {
  hour: string;
  workers: number;
  violations: number;
}

export interface ViolationCategory {
  category: string;
  count: number;
  color: string;
}

export type DetectedWorkerStatus =
  (typeof DetectedWorkerStatus)[keyof typeof DetectedWorkerStatus];

export const DetectedWorkerStatus = {
  safe: "safe",
  unsafe: "unsafe",
} as const;

export interface DetectedWorker {
  id: number;
  bbox: number[];
  status: DetectedWorkerStatus;
  confidence: number;
  violations: string[];
}

export interface DetectionSummary {
  total_workers: number;
  safe_workers: number;
  unsafe_workers: number;
  violations_detected: number;
}

export interface DetectionResult {
  /** @nullable */
  file_name?: string | null;
  file_size_kb: number;
  content_type: string;
  workers: DetectedWorker[];
  summary: DetectionSummary;
  model: string;
  note: string;
}

export interface AnalyticsSummary {
  totalWorkers: number;
  totalDetections: number;
  safeDetections: number;
  unsafeDetections: number;
  safetyScore: number;
}

export interface TrendDataPoint {
  day: string;
  total: number;
  unsafe: number;
  violations: number;
}

export type WorkerStatTopViolationsItem = {
  name: string;
  count: number;
};

export interface WorkerStat {
  totalDetections: number;
  safeDetections: number;
  unsafeDetections: number;
  safetyScore: number;
  topViolations: WorkerStatTopViolationsItem[];
}

export interface WorkerRecord {
  id: number;
  trackId: number;
  /** @nullable */
  name?: string | null;
  createdAt: string;
}

export type DetectionRecordStatus =
  (typeof DetectionRecordStatus)[keyof typeof DetectionRecordStatus];

export const DetectionRecordStatus = {
  safe: "safe",
  unsafe: "unsafe",
} as const;

export interface DetectionRecord {
  id: number;
  workerId: number;
  /** @nullable */
  workerName?: string | null;
  /** @nullable */
  trackId?: number | null;
  uploadId: string;
  status: DetectionRecordStatus;
  violations: string[];
  confidence: number;
  bbox: number[];
  timestamp: string;
}

export interface WorkerAnalytics {
  worker: WorkerRecord;
  stats: WorkerStat;
  recentDetections: DetectionRecord[];
}

export type DetectionHistoryPagePagination = {
  page: number;
  limit: number;
  total: number;
  totalPages: number;
};

export interface DetectionHistoryPage {
  data: DetectionRecord[];
  pagination: DetectionHistoryPagePagination;
}

export interface ErrorResponse {
  error: string;
}

export type DetectWorkersBody = {
  file: Blob;
};

export type GetDetectionHistoryParams = {
  page?: number;
  limit?: number;
  workerId?: number;
  status?: GetDetectionHistoryStatus;
  dateFrom?: string;
  dateTo?: string;
};

export type GetDetectionHistoryStatus =
  (typeof GetDetectionHistoryStatus)[keyof typeof GetDetectionHistoryStatus];

export const GetDetectionHistoryStatus = {
  safe: "safe",
  unsafe: "unsafe",
} as const;