# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Worker Safety Monitoring API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /detect:
    post:
      operationId: detectWorkers
      tags: [ai]
      summary: Detect workers in uploaded file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Detection results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectionResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /history:
    get:
      operationId: getDetectionHistory
      tags: [history]
      summary: Detection history with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: workerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [safe, unsafe]
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Paginated detection history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectionHistoryPage"
components:
  schemas:
...
    DetectedWorker:
      type: object
      properties:
        id:
          type: integer
        bbox:
          type: array
          items:
            type: integer
        status:
          type: string
          enum: [safe, unsafe]
        confidence:
          type: number
        violations:
          type: array
          items:
            type: string
      required: [id, bbox, status, confidence, violations]
    DetectionSummary:
      type: object
      properties:
        total_workers:
          type: integer
        safe_workers:
          type: integer
        unsafe_workers:
          type: integer
        violations_detected:
          type: integer
      required:
        - total_workers
        - safe_workers
        - unsafe_workers
        - violations_detected
    DetectionResult:
      type: object
      properties:
        file_name:
          type: ["string", "null"]
        file_size_kb:
          type: number
        content_type:
          type: string
        workers:
          type: array
          items:
            $ref: "#/components/schemas/DetectedWorker"
        summary:
          $ref: "#/components/schemas/DetectionSummary"
        model:
          type: string
        note:
          type: string
      required: [file_size_kb, content_type, workers, summary, model, note]
...
    DetectionRecord:
      type: object
      properties:
        id:
          type: integer
        workerId:
          type: integer
        workerName:
          type: ["string", "null"]
        trackId:
          type: ["integer", "null"]
        uploadId:
          type: string
        status:
          type: string
          enum: [safe, unsafe]
        violations:
          type: array
          items:
            type: string
        confidence:
          type: number
        bbox:
          type: array
          items:
            type: number
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - workerId
        - uploadId
        - status
        - violations
        - confidence
        - bbox
        - timestamp
    DetectionHistoryPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DetectionRecord"
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
          required: [page, limit, total, totalPages]
      required: [data, pagination]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]