# lib/api-zod/src/generated/types/detectedWorker.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Worker Safety Monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { DetectedWorkerStatus } from "./detectedWorkerStatus";

export interface DetectedWorker {
  id: number;
  bbox: number[];
  status: DetectedWorkerStatus;
  confidence: number;
  violations: string[];
}