# lib/api-zod/src/generated/types/detectionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Worker Safety Monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { DetectedWorker } from "./detectedWorker";
import type { DetectionSummary } from "./detectionSummary";

export interface DetectionResult {
  /** @nullable */
  file_name?: string | null;
  file_size_kb: number;
  content_type: string;
  workers: DetectedWorker[];
  summary: DetectionSummary;
  model: string;
  note: string;
}