# lib/api-zod/src/generated/types/detectionSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Worker Safety Monitoring API
 * OpenAPI spec version: 0.1.0
 */

export interface DetectionSummary {
  total_workers: number;
  safe_workers: number;
  unsafe_workers: number;
  violations_detected: number;
}